Fix AppReproducersTest#timezonesBakedIn
With GraalVM prior to 24.2.0 the default locale (and included locale)
for a native image was determined by the -Duser.language and
-Duser.country settings as image build time. With GraalVM >= 24.2.0 the
build settings of language/country have no effect in terms of included
locales in the resulting native image. Therefore, -H:IncludeLocales,
needs to be used at build time instead. Also at image runtime the
default language is determined by the system locale. For the purpose of
the test we'd want the French language which we need to set at image
runtime when the test runs.

Closes: #285