Fix vthreads app test when JAVA_HOME points to JDK 17
In some scenarios the Java build is using JDK 17 (default for quarkus),
but the vthreads integration test doesn't work on a Java version lower
than 21. Since these tests are native integration tests and the test
executes on Mandrel 24.2+ (JDK 24+) only so far, we can be sure that the
javac from the GraalVM/Mandrel installation supports virtual threads.
Use that over the one set in JAVA_HOME. Derive the value from where
native-image is located.

The container case is being handled elsewhere. Skip the test for
Windows for now.

Closes: #338