Add test for certificates
The JDK includes a cacerts file for establishing which root TLS
certificates it is going to trust. What's more it has a notion of
blocked certificates which should never be trusted. This patch adds a
check for both those things for native-image as the way those are
handled are different there.

Closes: #350