Increase timeout for backtrace command in debugSymbolsQuarkus
Using the existing `increasedTimeoutMS` which is set to 80000ms.  This
might be appear as too high, but we already use it for other commands
taking 10-15 secs, like the first breakpoint command, as we have seen
some pretty slow VMs hitting the timeout if it's not high enough.

Reusing an existing value also keeps things simpler.

Closes https://github.com/Karm/mandrel-integration-tests/issues/360