Allow list new warnings message on JDK 25+
Previously this was done on JDK 26+, but with the recent
reversion of graal/master being JDK 25-based again we don't
get the allow-listed warnings matched anymore.

Instead of doing a new version parsing scheme (again!) we just match the
version 25+ as it should do no harm to match this on Mandrel 25 as well.

Closes: #369